refactor: add login-buttons component and update translations
A new component "login-buttons" has been seperated to the UI library from the navbar user controls.
The relevant translations (in 'bg.json' and 'en.json') have been updated accordingly.
