+++ libs/sugaming-ui/src/lib/components/site/server.ts
@@ -1,2 +1,3 @@
 export * from './team-card/team-card';
+export * from './login-buttons/login-buttons';
 export * from './logo/logo';
